micr_reader/requirements.txt
# requirements.txt
# Dépendances pour le MICR Reader

# API OpenAI pour GPT-4o
openai>=1.0.0

# Traitement d'images
Pillow>=10.0.0

# Interface utilisateur
gradio>=4.0.0
pandas>=1.5.0

# Configuration et utilitaires
python-dotenv>=1.0.0

# Tests (optionnel)
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools (optionnel)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation (optionnel)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
